added try-catch in controllers

+++ controller/image.controller.js
@@ -8,28 +8,47 @@ class PostController {
             return;
         }
         const posts = await db.query('SELECT * FROM image WHERE person_id = $1 OR post_id = $2', [user_id, post_id]);
-        res.json(posts?.rows ?? []);
+        res.status(200).json(posts?.rows ?? []);
     };
 
     getOneImage = async (req, res) => {
         const { id } = req.params;
-        const posts = await db.query('SELECT * FROM post WHERE id = $1', [id]);
-        res.json(posts?.rows ?? []);
+
+        try {
+            if (!id) throw new Error('Invalid data');
+
+            const posts = await db.query('SELECT * FROM post WHERE id = $1', [id]);
+            res.status(200).json(posts?.rows ?? []);
+        } catch (ex) {
+            res.status(400).json({ Error: ex });
+        }
     };
     createImage = async (req, res) => {
         const { file_name, file_path, person_id, post_id } = req.body;
 
-        const newImage = await db.query(
-            'INSERT INTO image (file_name, file_path, person_id, post_id) VALUES ($1, $2, $3, $4) RETURNING *',
-            [file_name, file_path, person_id, post_id]
-        );
+        try {
+            if (!file_name || !file_path || !person_id || !post_id) throw new Error('Invalid data');
 
-        res.json(newImage?.rows?.[0]);
+            const newImage = await db.query(
+                'INSERT INTO image (file_name, file_path, person_id, post_id) VALUES ($1, $2, $3, $4) RETURNING *',
+                [file_name, file_path, person_id, post_id]
+            );
+            res.status(200).json(newImage?.rows?.[0]);
+        } catch (ex) {
+            res.status(400).json({ Error: ex });
+        }
     };
     deleteImage = async (req, res) => {
         const { id } = req.params;
-        const posts = await db.query('DELETE FROM image WHERE id = $1', [id]);
-        res.json(!!posts?.rowCount);
+
+        try {
+            if (!id) throw new Error('Invalid data');
+
+            const posts = await db.query('DELETE FROM image WHERE id = $1', [id]);
+            res.status(200).json(!!posts?.rowCount);
+        } catch (ex) {
+            res.status(400).json({ Error: ex });
+        }
     };
 }
